# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Retail Tractors review-service API",
    "description": "API documentation for the Retail Tractors review-service."
  },
  "host": "localhost:3003",
  "basePath": "/reviews",
  "tags": [
    {
      "name": "Reviews",
      "description": "Operations related to reviews"
    }
  ],
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Insert token as: Bearer <token>"
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/": {
      "post": {
        "tags": [
          "Reviews"
        ],
        "summary": "Create a new review",
        "description": "Creates a new review for equipment.",
        "operationId": "createReview",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Review object that needs to be added",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Review"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Review created successfully",
            "schema": {
              "$ref": "#/definitions/Review"
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "get": {
        "tags": [
          "Reviews"
        ],
        "summary": "Get all reviews",
        "description": "Retrieves a list of all reviews.",
        "operationId": "getAllReviews",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number for pagination",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of items per page",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "startDate",
            "in": "query",
            "description": "Filter reviews starting from this date (ISO 8601 format)",
            "required": false,
            "type": "string",
            "format": "date-time"
          },
          {
            "name": "endDate",
            "in": "query",
            "description": "Filter reviews ending by this date (ISO 8601 format)",
            "required": false,
            "type": "string",
            "format": "date-time"
          },
          {
            "name": "status",
            "in": "query",
            "description": "Filter reviews by status",
            "required": false,
            "type": "string"
          } 
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Review"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/{id}/status": {
      "put": {
        "tags": [
          "Reviews"
        ],
        "summary": "Update review status",
        "description": "Updates the status of an existing review.",
        "operationId": "updateReviewStatus",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the review to update",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Review status update object",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string"
                },
                "notes": {
                  "type": "string"
                }
              },
              "required": [
                "status"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Review status updated successfully",
            "schema": {
              "$ref": "#/definitions/Review"
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Review not found"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/check-availability": {
      "get": {
        "tags": [
          "Reviews"
        ],
        "summary": "Check equipment availability",
        "description": "Checks the availability of equipment for a given date range.",
        "operationId": "checkAvailability",
        "parameters": [
          {
            "name": "equipmentId",
            "in": "query",
            "description": "ID of the equipment to check",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "startDate",
            "in": "query",
            "description": "Start date of the desired booking period (ISO 8601 format)",
            "required": true,
            "type": "string",
            "format": "date-time"
          },
          {
            "name": "endDate",
            "in": "query",
            "description": "End date of the desired booking period (ISO 8601 format)",
            "required": true,
            "type": "string",
            "format": "date-time"
          }
        ],
        "responses": {
          "200": {
            "description": "Availability status",
            "schema": {
              "type": "object",
              "properties": {
                "available": {
                  "type": "boolean"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/{id}": {
      "delete": {
        "tags": [
          "Reviews"
        ],
        "summary": "Delete a review by ID",
        "description": "Deletes an existing review by ID. Admin only.",
        "operationId": "deleteReviewById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the review to delete",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "204": {
            "description": "Booking deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Booking not found"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "get": {
        "tags": [
          "Reviews"
        ],
        "summary": "Get all reviews",
        "description": "Retrieves a review by its ID.",
        "operationId": "getReviewById ",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the review to retrieve",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "$ref": "#/definitions/Review"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Review not found"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/user/{userId}": {
      "get": {
        "tags": [
          "Reviews"
        ],
        "summary": "Get reviews by user ID",
        "description": "Retrieves reviews for a specific user.",
        "operationId": "getReviewsByUserId",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "ID of the user",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "page",
            "in": "query",
            "description": "Page number for pagination",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of items per page",
            "required": false,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Review"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/equipment/{equipmentId}": {
      "get": {
        "tags": [
          "Reviews"
        ],
        "summary": "Get reviews by equipment ID",
        "description": "Retrieves reviews for a specific equipment.",
        "operationId": "getReviewsByEquipmentId",
        "parameters": [
          {
            "name": "equipmentId",
            "in": "path",
            "description": "ID of the equipment",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "page",
            "in": "query",
            "description": "Page number for pagination",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of items per page",
            "required": false,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Review "
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    }
  },
  "definitions": {
    "Booking": {
      "type": "object",
      "required": [
        "userId",
        "equipmentId",
        "startDate",
        "endDate"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "userId": {
          "type": "integer",
          "format": "int64"
        },
        "equipmentId": {
          "type": "integer",
          "format": "int64"
        },
        "startDate": {
          "type": "string",
          "format": "date-time"
        },
        "endDate": {
          "type": "string",
          "format": "date-time"
        },
        "status": {
          "type": "string"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        }
      }
    }
  }
}